Revert changes to double for wheel in dmsdk hid

+++ engine/hid/src/dmsdk/hid/hid.h
@@ -464,7 +464,7 @@ namespace dmHID
     {
         int32_t  m_PositionX;
         int32_t  m_PositionY;
-        double   m_Wheel;
+        int32_t  m_Wheel;
         uint32_t m_Buttons[MAX_MOUSE_BUTTON_COUNT / 32 + 1];
     };
